refactor: rename folder "essay" to "main"
