+++ src/lib/main/BlocksRenderer.svelte
