+++ src/lib/main/IllustrationDesktop.svelte
