+++ src/lib/main/IllustrationMobile.svelte
